re #39: Rework loading of resources. Add level attribute to capabilities and requirements to speed-up entities tree structure. BND files cleanup.

+++ crce-metadata-dao-impl/src/main/java/cz/zcu/kiv/crce/metadata/dao/internal/type/DbOperator.java
@@ -36,7 +36,7 @@ public enum DbOperator {
         throw new IllegalArgumentException("Invalid operator: " + operator);
     }
 
-    public static Operator getOperatorValue(int dbValue) {
+    public static Operator getOperatorValue(short dbValue) {
         for (DbOperator value : values()) {
             if (value.dbValue == dbValue) {
                 return value.operator;
